refactor:update src/api for split hello and config-xxx api path
