+++ src/api/v0/hello.py
@@ -1,0 +1,19 @@
+# hello.py
+from fastapi import APIRouter, Request
+from functions.AppLogger import AppLogger
+
+
+APP_NAME = "api_server"
+router = APIRouter(tags=["Config"])
+
+
+@router.get("/hello")
+@router.post("/hello")
+async def hello(request: Request):
+    """
+    GETとPOSTメソッドで`/api/v0/hello`エンドポイントにアクセスすると、
+    JSON形式で`{"result": "hello"}`を返します。
+    """
+    api_logger = AppLogger(APP_NAME)
+    api_logger.info_log(f"{request.url.path} Receive {request.method}")
+    return {"result": "hello"}
